// src/app/Models/task.ts
export class Task {
    Task1: string;
    ParentTask: string;
    StartDate: Date;
    EndDate: Date;
    Priority: number;
    ParentTaskId: number;
    Task_ID: number;
    Parent_ID: number;
    Start_Date: String;
    End_Date: String;
    IsTaskEnded : number;
}


export const taskData: any=""; 